requirements.txt
psycopg[binary]>=3.1
python-dotenv
typer[all]
matplotlib
plotly
pandas
psutil
pytest
requests
faker
questionary
python-socketio[client]
aiohttp